Tree changes to add flight nodes, construction
* Now that we're not using Construction for Hydrolox, I reinserted the
construction line. Now you have to do basic materials research before
applying it to engines or capsules.
* General Construction necessary for TL2 engines, second-gen capsules
* Adv Construction necessary for Staged Combustion, third-gen capsules
* TODO: Heat shields should go in Construction nodes.
* Add more flight nodes: High Speed Flight (was Aerodynamics) for Mach2
jets and the X-15, Advanced Jet Engines (was Supersonic) for Mach 3+
flight and for turbofans, Prototype Spaceplanes (was High Alt Flight)
for Dynasoar and the like, and Usable Spaceplanes (was advAero) for
actually-usable spaceplanes.