Avionics: turn off 'hide if ok' when press Alt-I
Normally, Alt-I will toggle visibility. However, if 'hide if ok' is
turned on, Alt-I will not make the GUI visible if controls are unlocked.
This change will reset 'hide if ok' to 'always show' if the GUI is
currently enabled but (due to 'hide if ok') hidden, so Alt-I really will
show it again.