Priced SXT elevons
Larger two are possibly/probably not totally accurate. The leading edges clipped
different amounts into the base I was using for comparisons, and there are some
anti-shock bodies (Or at least I think they're called that) that may or may not
have an additional effect, esp. with nuFAR. Tried to compensate for the former,
no clue about the latter.