Trying to find a place for DockPorts
Not exactly sure what I'm doing here, NathanKell, all I want is to be able to see the stock docking ports in my game. I don't care if they come way late in the tech tree, I just would like them in my game. It's either editing this, or editing Tree.cfg in RP-0 and blocking it from being re-written (which I'm not sure works or not, haven't tested). I was following a post made by NathanKell on Realism Overhaul thread about the absence of stock docking nodes in the RP-0 tech tree.