Update tree.yml for issue #417
These are the bulk of the tech node changes discussed in issue #417.  In a few places I've also altered Cost and/or EntryCost so that duplicate parts match (i.e. SRBs from ATK have the same costs as their KW Rocketry equivalent).

Also included are tech node placement for the new parts discussed in issue #417.  This includes the AJ10Adv (118F/118K), the Atlas II tank, and the Thor/Delta avionics parts.