Add avionics tech requirements
Based on comments in issue #384 which allow avionics module to have it's own tech requirements, I've added avionic tech to all of the appropriate Titan, Atlas and Centaur tanks.  I've tried to keep the avionics TL matching the construction TL.  Exceptions are the avionics for the Atlas SLV-3 which I put in the 1/2 step "Stability" node, and the avionics for the SLV-3C which I put in Specialized Control which appears to be the same tech level (and possibly the same node though I do see two separate nodes) as avionicsTL4.