Station contract fix
According to the Contract Configuration documentation:
If this Rendezvous parameter is a child of a VesselParameterGroup parameter, then no more than *one* vessel should be provided (the other is the vessel being tracked under the VesselParameterGroup). If no vessel attributes are provided, the second vessel will match any vessel.

Since "Rendezvous" is a PARAMETER of "VesselParameterGroup", then the second "vessel" value is not needed and appears to be disabling the contract.