Place new Centaur tank
In RO Github 1126 I'm proposing adding FASAGeminiLFTCentarCSM_D1T to represent the insulated Centaur D-1T used on Titan 3E.  Witht he exception of the added thermal insulation, it is basically identical to the standard FASAGeminiLFTCentar so I put it in the same node and added a slight increase in cost to account for the extra insulation.

In the same Github I'm suggesting removing FASAGeminiLFTCentarCSM_D3 since Atlas IIIB did not use it's own version of the Centaur.  I've therefore removed it's entry from the tree.