Match RO-stock & FASA LR-105 entry costs
Long explanation for this. I am studying the related unlock costs of the LR89, LR79, and LR105 engines.

The smallest issue I noticed was this one; that one LR105 had a maxEntryCost and the other didn't. The initial LR105 entry cost is 11k if neither of the other two are unlocked. When the LR89 is bought (EntryCost paid) the FASA version of the 105 drops to a 3k entry price, while the RO was at 6k. After also unlocking the LR79, FASA 105 drops to 2k, while the RO version stays the same. Removing this line aligns them.

An aside. For future planning, should/could these unlock prices be linked to the global engine config, rather than to the part name? That may not be useful, as we may be phasing out the RO clone parts in 1.1 as we can just use a good model (ie. Raidernick ones) as the default. However, this idea may still be useful as it still may make it easier to expand support for other mods that include their own versions of engines we already have configured. In theory, using the Global Engine Config tag to prompt a script to set entry prices for all parts with that tag would save a lot of configing. Ideas? @SirKeplan @NathanKell