SSTU MUS Tanks
Right now, if you use one of the SSTU MUS tanks, you have unlimited avionics control.  Difficulty is, these tanks are procedural so the larger the tank, the more avionics you'd need for it.  What I've done here isn't pretty since the parts will now display 9 separate avionics modules, but in VAB/SPH and in flight, they do combine correctly.  So if, for example,  you have Flight Control, the MUS tanks can handle up to 40t for avionics purposes.  This doesn't have anything to do with the actual size of the part you create, but it does allow these tanks to function better throughout the course of a career game.  If there is a cleaner way to setup something like this, that would be a better option, but this at least works.