fix perl shebang line
use `/usr/bin/env perl` to look it up in the PATH instead of
the hardcoded /usr/bin/perl (typically the first perl in the user's
PATH is going to be preferred).

the -w flag causes issues when used with env but both of these
scripts declare `use warnings` so that flag is redundant and
removed from the shebang.

Signed-off-by: Lamont Granquist <lamont@scriptkiddie.org>