MAJOR SCIENCE OVERHAUL
***SAVE BREAKING***
******YOU MUST HAVE DMAGIC ORBITAL SCIENCE INSTALLED******

* Every aspect of the science system has changed
* There are new experiments, new transmission rates, new locations where
experiments can be run, and new tier levels to many experiments

* TIER SYSTEM *
* As in real life, technology gets better through time. The system
reflects this by having new parts that provide science on the same
experiments.
* The descriptions of each part have the science rates available based
on the Level of the part
* Most experiments are broken out into 4 levels that will improve with
technology

* More information will be forthcoming. This is the first part of the
major science update. Coming soon are new Crew Experiments for the
Advanced Capsules and Science Labs as well as rover based science
experiments
* I will be adding information to the Wiki for players to consult