Fix NewCrewStations.cfg
As I read this, the point of this contract is to send crew to an empty station.  But the current requirement is set to only allow the contract to occur when there are less than 3 vacant seats available.  Meaning if a station can support 5 crew and you have 3 on the station, the contract would become available because "EmptyCrewSpace()" would return a result of 2.  By using "v.CrewCount()==0" we should only allow the contract to occur when a station has no crew on board, regardless of how many available seats there may be.