Slinq-ify ALL the LINQs
Hopefully keeps the LINQ style while avoiding garbage.

Some of the functions take two-arg lambdas and an additional argument to
avoid having to allocate to capture an external variable. This results
in some pretty bad variable names, unfortunately.

Where(x => <cond>).FirstOrDefault() can be replaced with
FirstOrDefault(x => <cond>) for simple lambdas. Unfortunately, Slinq
doesn't provide a FirstOrDefault() implementation that takes a two-arg
lambda and an additional argument, so Where(x => <cond>).FirstOrDefault()
is used to avoid allocations.

In one case, a LINQ statement was replaced with a function call, since
the LINQ statement was exactly the same as that inside the called
function.

Ended up replacing a FirstOrDefault() with First() since the code
accessed a member variable without checking for null first, so an empty
result would result in an exception either way. First() makes it a bit
clearer that something is expected, though.

Also clean up using statements a bit.