Merge Master to Developmental (#939)
* Tell the player Lunar Rover can be done 3 times (#926)

maxCompletions was bumped up from 2 to 3 a while ago, but description still claimed 2. A nice surprise, but accuracy is better.

* Small Cleanup

* Move SSTU Tank Diameter Upgrades from RO

They belong in RP-1 instead of RO because they affect technology unlocks

* Update RP-0.netkan

This fixes the RP-0/RP-1.netkan issue

* dead link (#933)

update forum link

* Update CustomBarnKit.cfg

Change the Launch Pad so that by Level 3 you are using the largest one.

* Update RP-0.netkan

Removed the B9 Procedural Wings Version that was on CKAN as it is not working.