Fixed the names of all of the parts
credit to Kerrigan778 on discord for spotting the issue and putting together the bulk of this fix. 

Basically all of the names for each part as well as the folder structure appears to have changed since the last implementation of this config, all this does is fixes the names to the actual ones so that the parts will work properly.