RN Stations balancing
Balanced the pricing of the Skylab and Salyut stations.
The cost of the whole station stayed about the same, but now the actual station part makes up more of the cost.
Entry Cost went way up with no previous parts unlocked, but down with some reasonable tech (ECMs) "already owned