Unlock First EVA with any EVA-enabling tech
First EVA milestone was recently hidden behind 'must unlock 2nd gen
capsules', to protect players from thinking they can do it with
Mercury

But 2ndgen capsules aren't the only way to go EVA, prototype spaceplanes
are an alternate path. It'd be silly if the player could go EVA and not
hit the milestone.