Move hydrazine back a node
Hand-in-hand with https://github.com/KSP-RO/RP-0/pull/1240.

Once Cavea-B is moved back, the "stability" node will be nearly empty. NASA document 19680006875 says that hydrazine monoprop was fired in space in 59 as a test but not used operationally until 1961 with Ranger and Mariner. This moves it from the '59 node to the '61 node.