Change upgrade costs and tiers of R&D and Mission Control
In the R&D building, upgrade costs are pushed back 1 tier,
so that maximum cost is 6M instead of 10M, and every tier has lower cost
Additionally, levels 3 and 4 now have higher node cost limits.

In the Mission Control, more levels are added, with 5 instead of 3,
and upgrade costs are overall reduced.