Fix hiring Hero kerbals after they would have retired (#1410)
* Force retiring kerbals to be crew
KSP 1.10 introduced a change which causes the hero kerbals to return to the applicant list when fired. If they are then re-hired after their retirement date, they immediately retire.
In order to prevent re-hiring once they've retired, we can force kerbals to become "crew" when their retirement date hits. This will properly move them out of the applicant pool and into the "lost" list.

* Prevent pushing back retirement by firing a kerbel
Previously, if a hero kerbal was fired and then re-hired, their retirement date would be reset.
Now we only update retirement date if  a kerbal is brand new