Unlock the first tank configs from the start
So that Experimental Parts Int and Balloon tanks will have a
valid tank config, instead of a confusing tank UI that says
no fuels are valid.

Actual tank parts remain restricted by the tech tree, so this
shouldn't allow building anything that shouldn't be.

Should also makes the behavior when loading a pre-existing .craft
with a not-yet-unlocked config marginally more sane.