Add Engine Refubishment
This discounts the cost of used engines so that when replaced with new ones in the editor there will be a charge.   Formula is stored as a KCT formula. This is just proof of concept, needs to be updated with new changes for TestFlight's cumulative runtime feature and at minimum expanded to ablators.