Add vessel merging and engine refurbishment cost. (#1551)
Transcribed changes from
https://github.com/linuxgurugamer/KCT/pull/76 and
https://github.com/linuxgurugamer/KCT/pull/83

* Add Human Rated Flag

Adds 0 cost Human Rated Part to 1961 command pods node.  Primary use is to flag LVs as human rated for later joining with a human rated stage.

* Add Engine Refubishment

This discounts the cost of used engines so that when replaced with new ones in the editor there will be a charge.   Formula is stored as a KCT formula. This is just proof of concept, needs to be updated with new changes for TestFlight's cumulative runtime feature and at minimum expanded to ablators.