Fix cost of ROE-215, RN RD-216
They both use a 2x engine multiplier, but seem to be costed for the
1x rd-215.

This change makes the initial config an ok deal at 700 and the rd215m
still a decent deal at 600 (lr87 goes 500-760-...) instead of a great
deal and a fantastic deal at 350/250.

Other configs (rd217/225/250) are trickier to compare, but at least
probably aren't becoming overpriced to the point of uselessness.