Improve upgrade Descriptions
Clarify that you don't get the part that serves as the icon for airlaunch upgrades.
Minor improvements to airlaunch upgrade descriptions.
Add descriptions for each comm tech level from the Real Antennas wiki.
Add descriptions to tanks showing the difference in weight and cost relative to the previous level.
Clarify that MLI upgrades must be purchased.
Include in description for first MLI upgrade that they cannot be applied to balloon tanks at that tech level