Add contracts to the Earth Space Station Program (#1835)
Adds a set of 3 contracts based on the Skylab station. 1st contract requires launching a more advanced station and crewing it with 3 crew for 30d. 2nd contract involves crewing for 60d, and 3rd contract involves crewing for 90d. Both 2nd & 3rd contract can be completed twice, with the former intended to be required and the latter optional.