Give Early Film Camera all the slots
22 slots covers 11 earth biomes, flying high + space low.

Because it's silly that a camera can't take pictures of trees after having
taken pictures of water, grass and rocks; as long as there's film, let it
take pictures.

Also because juggling slots is too micromanagey.
There will still be micromanagement if the camera isn't recovered and the
recovery part has too few slots, but this won't make it any worse?