Ensure that reconstruct does not downsize the hangar
Unlike in LCs, the hangar does not pay maintenance for increases to maximum dimensions, and since they dictate maximum engineer count it's important not to downsize it unnecessarily. This is particularly so because many planes can use several times more engineers than a hangar built to their dimensions supports--presently pressing "reconstruct" can cripple your ability to build the plane selected during reconstruction.

I've tested that this change provides what I believe to be the correct behavior--hangar reconstruction and upgrades upsize but do not downsize, while LC reconstruction downsizes but upgrades do not.