Update ModuleShowInfo
The `name` field on a `Part` is not always the part name; in particular
when creating the tooltip in the editor for a part with upgrades, it is
not (it is instead "{partName}(Clone)").

Fixes an issue where the required crew training course name in the
tooltip does not match the course name in the RP-1 training window.