Increase Tundra orbit periapsis requirement (#2375)
With a 1000 km periapsis (eccentricity 0.82), the resulting orbit has
barely any apogee dwell time; it moves west very quickly across the sky
when it's at apogee.

At an eccentricity of 0.423, the ground track moves from being a
figure-of-eight to a teardrop shape. As far as I can tell, this is the
highest eccentricity that gives a proper Tundra orbit.

A 17000 km periapsis corresponds to an eccentricity of around 0.446,
which makes the maximum eccentricity a little higher than that value.

Also decreased the minimum eccentricity a little bit. According to
Capderou's "Handbook of Satellite Orbits", a 0.2668 eccentricity Tundra
orbit was considered for ESA's "Archimedes" telecommunications project.