Fix a timewarp kraken, add MNA and ArgPe to simulation GUI options, and some other various simulation GUI changes (#2494)
* tell user that they can input seconds in the simulation datetime

since DateTime can ofc handle seconds

* avoid timewarp kraken

* also remind users of allowing time in seconds

who would use this?

* Update GUI_Simulation.cs

* Update GUI_Simulation.cs

* prevent user from inputting exact date and time when "From Now" is selected

* Update GUI_Simulation.cs

* add small buffer to default sim orbit altitude

especially a problem with principia

* add mean anomaly and argument of periapsis to simulation gui

is that some good code siimav? ill be stealing that, thank you very much

* Update SimulationParams.cs

* Update SpaceCenterManagement.cs

* Update SimulationParams.cs

* change ArgPE to ArgPe

its argPe in the docs so this is a little better

* misc changes

* fix logic

previously, it was doing:
if (!whitespace and (contains ":" and does not have date format)) OR fails parse

now its doing:
if !whitespace and ((contains ":" and does not have date format) OR fails parse)

this should correctly allow for whitespaces

* cut tip text down, as it was overrunning the box

* change default height above atmosphere orbit to be 5000 instead of 20000

if we already tell the user what the minimum value is, then we dont need to set them so high above the atmosphere, as this could just be confusing. 5000 is more than high enough to counteract principia shenanigans, while still allowing the user to be in a low orbit by default

* change position of "(s)" in delay, add "km" to min and max

* increase width of info box?

does this even work?

* revert previous change

just increases the size of the simulation config box, not the tooltip