Fixes issue with Kerbalism Planner reporting high EC consumption in Editor because of an unknown "lab" process. (#2624)
https://discord.com/channels/418222516386004992/418222516386004994/1408348148023300109
Adding %MODULE[ModuleScienceConverter] needs to happen before/above the patch that will look for all parts that have that module so it can replace it.
Since it is currently at the bottom, it will add a blank module with no info and I guess Kerbalism is just determining that to be the max that a "lab" will consume.